Fixed #14 -- Allow customizing the redirect query param name
Allow a page to invoke a login followed by a redirect to the given page. This
mirrors the functionality of the default login view.

It also brings over the login view's `redirect_field_name` parameter,
allowing a site to customize what query parameter to use. This helps
when transitioning older codebases to the new view.

The template by default uses "next", but this can be overridden in
custom templates.

This adds unit tests covering a custom redirect URL using the standard
`?next=` query parameter and a custom query parameter defined by
`redirect_field_name`.

Thanks to Christian Hammond for the patch.