Provide the redirect URL in the login form, and allow custom field names.
This provides the redirect URL provided in the query string (generally
"next", as per REDIRECT_FIELD_NAME) to the login form, allowing a page
to invoke a login followed by a redirect to the given page. This mirrors
the functionality of the default login view.

It also brings over the login view's `redirect_field_name` parameter,
allowing a site to customize what query parameter to use. This helps
when transitioning older codebases to the new view.

The template by default uses "next", but this can be overridden in
custom templates.