Add unit tests, and remove a useless <input/> in the form.
This adds unit tests covering a custom redirect URL using the standard
`?next=` query parameter and a custom query parameter defined by
`redirect_field_name`.

It also removes the <input/> tag I added to the template in the original
commit, as it's not actually useful, given that the form will continue
to pass the redirect URL in the GET query string.