Make the "Back to profile" label consistent with other buttons.
The "Back to profile" button label has been changed to "Back to Profile"
in order to be consistent with other button title casings. Title Casing
is fairly standard for most button labels, except when in
full question/sentence format, and is a good fit for most sites. This
was the only button label not asking a question that wasn't consistent.