Updated Travis CI / tox build matrix
Included tests for YubiKey support.

On Travis CI a simplified build matrix is used, covering all Python versions
and all Django versions, but only for the most common combinations of the two.

On tox, a full cross-check with all supported versions and options is
performed. Hint: use detox to run the tests in parallel.