Fix #182: Autofocus token input during authentication
Token input usually comes directly after user:pass input. The user
already has their hands on the keyboard and the token is the only field
for input in the wizard flow. It should auto focus, so no additional tab
or mouse clicks are necessary to complete authentication.

While this could be a template decision, since the authentication
workflow is wrapped in a wizard, I think changing the
AuthenticationTokenForm makes the most sense.