Move login redirect URL logic into its own method
Move the determination of the redirect URL to its own method, when redirecting in LoginView.done. This makes it possible for users to make their own class where `get_redirect_url` is overridden, so that business logic can determine the next URL. For example, [Zulip](https://github.com/zulip/zulip/blob/master/zerver/views/auth.py#L700-L708) has user-specific "realms", so they want to determine the redirect URL only after login.

This moves the redirect URL logic to `get_redirect_url`, and calls that just before calling `redirect`. This moves the URL logic after `signals.user_verified.send`, which means a change in behavior if the URL logic raises an exception. For example, if the given URL cannot be resolved. It used to not call `signals.user_verified.send`, but now it does. I think this is expected behavior; the user is logged in after all.

The default `get_redirect_url` shouldn't be able to redirect to an unsafe (off-domain) URL. When overriding `get_redirect_url` it can do whatever it wants, including redirecting to an unsafe URL. This commit adds tests to reflect this.