:hammer: Re-organize dependencies for local dev and CI
The 'hard' dependencies are now specified as extra groups in the
setup.py package information. tox.ini points to these extras to
install the required dependencies, and their supported version ranges.

The dev requirements file is now documented in the contributing
guidelines, and makes use of the extras to pull in the optional
dependencies for a full-blown local installation. The remaining
dependencies (such as sphinx) are not used in tox and are still
in the requirements_dev.txt file.

The linting dependencies (with upper bound versions) are now
also specified as an extra group to prevent mismatching
versions in tox env and local development envs.